MB-32661: [BP] Fix expiration decoding with docflags included
This patch fixes an issue where mutations would ignore and reset
expiry times back to 0 if docflags were included in the extras.
The test also included in this patch tests this, which is added
to an existing similar test to avoid adding more sleep time.

Change-Id: I10d8c766de220c2d149c9e007c33d95fe0710e65
Reviewed-on: http://review.couchbase.org/103668
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Daniel Owen <owend@couchbase.com>