MB-32618: Improve logging when inflate fails
Snappy decompression can fail for three reasons:
1) snappy::GetUncompressedLength returns false
2) inflated_length > max_inflated_size
3) snappy::RawUncompress returns false

For debugging purposes it is useful to know which of the reasons caused
the inflate to fail, particularly in the case we throw an exception.

The platform API for compression provides the ability to obtain the
uncompressed length (get_uncompressed_length).  Therefore log
the buffer size and uncompressed length (if possible) in the exception
message.

Change-Id: I179a944c4a6cbd134682cf8d17e85c35c3c14833
Reviewed-on: http://review.couchbase.org/103669
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>