MB-32669: Set datatype when reading an evicted value on expiry
The expiry path needs the value for xattr docs (deleted or not)
so issues a read to kvstore. If the hash-table metadata.datatype
was set to snappy by the active compressor the value passed to
the expiry hook has incorrect datatype and triggers a crash.

Change-Id: I994ef0e71bb6402b0954289a90125741f0bc75e3
Reviewed-on: http://review.couchbase.org/103670
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>