MB-32685: Delay updating RBAC db revision number
Use a dedicated counter to track the revision number
for the RBAC privilege database, and set the current
revision number _after_ it is installed.

This works around the problem where other clients
try to perform a privilege check at the same time
as the privilege database was built. Previously
the current number was bumped while building the new
database, which would cause clients to get AUTH_STALE
when they tried to check privileges. The code would
try to rebuild the privilege context, but it would
receive a context from the previous version and still
be stale until the thread rebuilding the privilege
database completed.

Change-Id: I54d1844a299e0373262737a3107641fd449b9560
Reviewed-on: http://review.couchbase.org/103729
Tested-by: Build Bot <build@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>