MB-32684 [SR]: Record setVBstate VBucket topology
Pass the new optional JSON payload for the SET_VBUCKET_STATE request
down to ep-engine's VBucket, validate it and record the specified
topology inside VBucket.

Expose the topology via a new 'topology' key in the 'vbucket-details'
stat group. For example, when injecting some dummy data:

    $ cbstats localhost:12000 -u Administrator -p asdasd vbucket-details 990 |grep topology
    vb_990:topology:                        [["ns_1@ABCDXYZ.ab1","ns_1@ABCDXYZ.cd1"],["ns_1@ABCDXYZ.ab5","ns_1@ABCDXYZ.cd6"]]

Note the topology is currently ignored by the DurabilityMonitor,
because ns_server is not yet populating this value and we have a
temporary hardcoded topology for testing. Once ns_server is sending
the topology and node name identifiers then it can be connected up and
the temporary hardcoded topology removed.

Change-Id: Ic8cbd9c81def766b707a6d4fccaac27f5c9031ff
Reviewed-on: http://review.couchbase.org/103740
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>