MB-32443: Introduce IS_EXPIRATION flag in DelWithMeta options
This patch adds a new flag inside DelWithMeta's options field that
indicates whether the delete with meta message is referring to an
expiration instead of a standard explicit delete. The deletionSource
can be inferred from the status of this flag, which allows XDCR and
other users of delete with meta to correctly retain the deletion
source of a document, fixing the linking issue for expiry opcode
output.

Change-Id: I0054ab612552ca2d403aa5bce705200d663ce745
Reviewed-on: http://review.couchbase.org/103736
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>