Refactor: Use strongly-typed enum for setVBucketState.transfer
Change the type of the transfer parameter of setVBucketState() from a
bool to a new explicit enum class. The primary reason for this is to
make it safer as we add more optional arguments to this function
(e.g. forthcoming optional JSON payload), as nlohmann::json can be
constructed with a bool (and hence care needs to be taken with
argument order).

Change-Id: Ie46115dc78dcb86d3cfe51f8affd6af3578e6082
Reviewed-on: http://review.couchbase.org/103744
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>