MB-32712: Move BuckerLogger message prefixing from sink_it_() to log()
Move the prefixing of connection id, bucket name etc. from the
'backend' sink_it_() method to the 'frontend' log() method.

This allows[1] the BucketLogger class to be completely "aware" of all
heap allocations made while logging a message - previously the
base-class 'spdlog::logger' could allocate heap memory in the log()
method before the BucketLogger class was aware of it (and hence before
we could switch away from the current engine).

This is necessary because the heap memory will ultimately be freed by
spdlog's background flusher thread (which is shared between all
buckets and hence isn't associated with any of them).

An alternative approach would be to switch away from the current
engine in the EP_LOG_xxx() macros which most callers use; but this
has two problems:

1. By the time we get to trying to prefix the buckets' name onto the
   log message (in sink_it_), we have already switched away from the
   current engine so don't know what engine name to print.

2. Not all log messages us the EP_LOG_xxx() macros - for example DCP
   ActiveStream objects create their own BucketLogger instances to be
   able to automatically prefix the message with the stream name.

[1] Note this patch doesn't actually fix the mis-accounting - memory
is still tracked in the same way as before - this is just a refactor
to allow that in a subsequent patch.

Change-Id: I0f08e1619281f44798dbdd59ba6e5edc32b99228
Reviewed-on: http://review.couchbase.org/103788
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>