MB-32591: Introduce ChkPtQueueIterator
To help reduce memory usage we want to be able to remove items from the
checkpoint queue where all cursors are to the right (i.e. the items have
been processed) and hence don't need to be held in the queue.

The first step in this is to wrap the checkpoint queue iterator, (an
iterator on a stl::list) in a special iterator that automatically
skips passed elements that contain null pointers.

Change-Id: Iff294489d66c480d106ab1647585265f2fc4e651
Reviewed-on: http://review.couchbase.org/103776
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>