MB-31266: Pass a isDropped(key) callback to ephemeral purger
The ephemeral purger will perform the same function as the
compactor in persistent buckets for the processing of dropped
collections.

All keys are checked to see if they are owned by a dropped
collection by a callback to the KVBucket::collectionsEraseKey
wired through to the purger.

Tests are now executed against persistent and ephmeral buckets
in all eviction modes.

Change-Id: Ice018f712fde73aa27ed0dbe9d5fcb543c4a1b02
Reviewed-on: http://review.couchbase.org/103785
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Chris Farman <christopher.farman@couchbase.com>