MB-32579: Set the TLS engine before deleting a VBucket
The MB highlighted a path where the use of shared_ptrs
meant that the VBucket was deleted on a path which
had set the TLE engine to null (ExecutorPool::doTaskStats).
This can lead to an assert if a VBucket is freed from within
that scope.

To address the issue

1) Set the TLS engine to the task's engine
2) Explicitly reset the vbucket (unique_ptr)
3) Explicitly force description to free by swapping with an
   empty string

There may still be an accounting issue as the GlobalTask itself
is freed from doTaskStats.

Change-Id: Ie75fc844c4417f7f6ffc6225192b18a920483f93
Reviewed-on: http://review.couchbase.org/103772
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>