MB-31266: EphemeralBucket::notifyNewSeqno must check vb is valid
When a setVBucket occurs, the new vbucket object is initialised with
the last collections manifest the node received, this will trigger
create/delete collection events which push data onto the checkpoint.

This event creation code will hit EphemeralBucket::notifyNewSeqno which
gets the VBucket pointer from vbmap but doesn't check it is valid,
causing a crash in the case we are still initialising the vbucket's
collection state.

Change-Id: Id7b208ec389abf90af164fcf6a3abf5e02d6c41e
Reviewed-on: http://review.couchbase.org/103879
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>