MB-32117: Stamp Chain into SyncWrite and enable disk-seqno tracking
In general, for supporting the wide set of Durability Requirements we
need to know what node has ack'ed a tracked SyncWrite. For achieving
that, we stamp the chain information into any new tracked SyncWrite.

That allows a more flexible Durability Requirements verification.
In this patch, the new verification is already used for enabling
disk-seqno tracking in the DurabilityMonitor.

Adding the chain-info in SyncWrite is also necessary for adding support
for Multiple-Replicas and Multiple-Replication-Chain, which will be
implemented shortly.

Change-Id: I10bb6ad5ba9dd6a9d04de1963f81c03b65cebde6
Reviewed-on: http://review.couchbase.org/103891
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>