MB-32762: Refactor: change ADD_STAT to std::function 'AddStatFn'
Allows greater flexibility in how ADD_STAT is implemented; not limited
to the only state being cookie.

This is used by subsequent patches to wrap ADD_STAT so its memory
usage is correctly tracked.

One downside of std::function compared to a raw function pointer is it
is more costly to copy. As such, also update all the uses of ADD_STAT
to pass as const reference so there's no additional runtime cost for
parameter passing. Given this results in touching virtually all uses
of ADD_STAT, rename it to match the current coding standards naming
convetion - AddStatFn.

Change-Id: I358833c7012652ac07f6efda73f4bf4a0d058294
Reviewed-on: http://review.couchbase.org/103893
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>