MB-32762: Refactor: change ADD_RESPONSE to std::function
Allows greater flexibility in how ADD_RESPONSE is implemented; not
limited to the only state being cookie.

This is used by subsequent patches to wrap ADD_RESPONSE so its memory
usage is correctly tracked.

As per ADD_STAT change, one downside of std::function compared to a
raw function pointer is it is more costly to copy. As such, also
update all the uses of ADD_RESPONSE to pass as const reference so
there's no additional runtime cost for parameter passing. Given this
results in touching virtually all uses of ADD_RESPONSE, rename it to
match the current coding standards naming convetion - AddResponseFn.

Change-Id: I94d88e1fc8e109bb3ed84477de7588cf96bc33e4
Reviewed-on: http://review.couchbase.org/103902
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>