MB-32361: Trigger collection erase after warmup
After a restart we need to check if any incomplete
collection drops need restarting, this requires
warmupComplete calling into the collections code
and then checking with each manifest to see if
a drop is in progress.

Note on backfill-manager.cc change:

This was required when fake_executorpool began
skipping dead tasks, the destructor assumed that
if a backfill was still in the active list then
the dcpconnmap would be live and a decrement could
occur. This is not the case, the connmap could of
destructed before the backfill-manager. Removing
the decrement should have no adverse affects as
we're in a shutdown path.

Change-Id: Ic16264e9ad1d256f74d3a785926da27b9fa9366d
Reviewed-on: http://review.couchbase.org/103963
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>