MB-32762: Wrap AddResponseFn in ExitBorderGuard thunk
The AddResponseFn callback may allocate memory (temporary buffers for
stat data) which will be de-allocated inside the server, after the
engine call (response) has returned. As such we do not want to account
such memory against this bucket.

Use makeExitBorderGuard to wrap the AddResponseFn callback. This ensures
that memory is correctly accounted if allocated in the context of
add_stat.

(Note: Previous code /appeared/ correct so this is functionally a
no-op, however it moves to a standard way of performing checks when we
exit ep-engine within a call to it.)

Change-Id: I11b3c0d783279c4225851a18ca2366d458867edd
Reviewed-on: http://review.couchbase.org/103903
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>