MB-32762: Wrap AddStatFn callback in ExitBorderGuard thunk
The AddStatFn callback may allocate memory (temporary buffers for stat
data) which will be de-allocated inside the server, after the engine
call (get_stat) has returned. As such we do not want to account such
memory against this bucket.

Create a variadic templare function `makeExitBorderGuard`, which can
be used to create thunks which will switch away from the current
engine before invoking the wraped callback; then return to the saved
engine on return.

Use makeExitBorderGuard to wrap the AddStatFn callback. This ensures
that memory is correctly accounted if allocated in the context of
add_stat.

(Note: Previous code /appeared/ correct so this is functionally a
no-op, however it moves to a standard way of performing checks when we
exit ep-engine within a call to it.)

Change-Id: I4c62687ec2c3f57f0a7d22f492c625f3b3038a54
Reviewed-on: http://review.couchbase.org/103894
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>