MB-32684 [SR]: Don't ignore vbstats change if non-empty meta
Historically KVBucket::setVBucketState() would short-circuit and
return success if the new state was the same as the old.

However with the addition of extra meta information for topology
changes, although the overall state may be the same the topology could
have changed.

In this case don't short-circuit if there's a non-empty meta.

Change-Id: I9d4b442f8c2e5b51851c42e83f81e081ce1149eb
Reviewed-on: http://review.couchbase.org/103904
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>