MB-32785: Replica acks memory-seqno at snapshot-end
Currently a Replica sends a SeqnoAck with the memory-seqno as soon as a
Prepare is received.
But, for achieving the consistency model that we aim to (consistency at
failure scenario, details in DesignDoc) the Replica cannot acknowledge
mutations in the middle of a snapshot.

With this patch, a Replica sends an aggregated SeqnoAck only when it
receives the snapshot-end mutation, and only if the snapshot contains at
least one durable write.

Change-Id: I9d433940d5ffdc8facddec48604eae4a8ea31ad1
Reviewed-on: http://review.couchbase.org/103964
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>