Minimise where statwriter.h is #include'd
statwriter.h is a somewhat expensive header to include - it contains a
number of templated functions which use <ostream> / <stream> and hence
including statwriter.h pulls in a large amount of code. Remove uses of
statwriter.h from other header files where possible to minimise the
cost.

In the case of ConnHandler, it has a template member function for
printing stat values of any printable type -
ConnHandler::addStat(). Move the definition of this function to a new
'connhandler_impl.h' - which in turn allows the removal of
statwriter.h from connhandler.h. The new connhandler_impl.h file
should be included by any files which need to actually call addStat -
e.g. ConnHandler and it's subclasses.

Change-Id: Ie34766b3862ee93f0c7b086d74497ead2089c623
Reviewed-on: http://review.couchbase.org/104068
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>