MB-32591: Expel duplicate entries in checkpoint queue
Instead of erasing the duplicate item pointers from the CheckpointQueue
we expel them (which means we keep a placeholder in the queue but
set the entry to null.)

This is in preparation for when we move the CheckpointQueue
implementation over from a std:list to a std:deque, and so cannot
efficiently erase individual items from the queue.

The CheckpointIterator is also updated to "behind-the-scenes" skip-over
these null elements, and hence operate the same as if the element
had been erased.

Change-Id: I2b41f69448f68e929c34c50817c4a95fff6d4dd9
Reviewed-on: http://review.couchbase.org/104082
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>