Move PermittedVBStates to its own file
PermittedVBStates is a simple helper class to represent a set of
allowed vbucket states. It is used in a few places inside KV-Engine,
but is currently declared inside <memcached/vbucket.h>. This requires
that <memcached/vbucket.h> includes the headers required by
PermittedVBStates - one of which (<platform/bitset.h>) is relatively
heavyweight.

Additionally PermittedVBStates is only used within ep-engine, so it
probably shouldn't live in a memcached/ header.

Move PermittedVBStates into it's own header file and update various
uses to minise the necessary #includes.

Change-Id: I40b3b591204989fe133769d73d27137df333ebc9
Reviewed-on: http://review.couchbase.org/104089
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>