JenkinsLogSearch: Add regex search functionality
This commit, along with some comment updates, allows multiple
search parameters to be added via the use of RegEx formatting
eg: -s (.*foo.*)(?:.*\n)*(.*bar.*)
The use of groups is preferred and in output formats, groups
will be printed out separately. If no groups are provided, the
whole RegEx will be used.

Change-Id: I463f9a5147349bcd80a72258dde8b4c7354ffede
Reviewed-on: http://review.couchbase.org/104095
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>