Refactor: Don't return error if the packet exceeds the max size
The server won't accept packets which exceeds a maximum treshold
which is set to 10MB higher than the current maximum document size
(which is 20MB).

Up until now we've returned with an error message for this packet,
but that has some disadvantagees in the server. We need to create
and initialize a cookie object and move to a temporary state
before closing the connection. Given that we currently gracefully
handle all packets up to 30MB we can silently close connections
trying to send even bigger packets.

Change-Id: Ic807e654fcdbff76b7a0e33f61404c3bc6e960e1
Reviewed-on: http://review.couchbase.org/103029
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>