MB-32734 [SR]: Add SyncDelete support to HashTable
Update HashTable::unlocked_softDelete() to take a SyncDelete argument
which specifies if the delete should be performed durably. Similar to
SyncWrites, this will stage an additional pending Item in the
HashTable with the proposed deleted state. Upon commit() this will
replace the existing item.

Expand HashTablePerspectiveTest and VBucketDurabilityTest to cover the
basic use-cases of this.

Change-Id: Iab4bb4c384bce6e7680cb523de96ed8cd71cdc00
Reviewed-on: http://review.couchbase.org/104152
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>