MB-32833 [SR]: Consumer sends its name as part of DCP negotiation
Introduce the new 'consumer_name' DCP_CONTROL key.
A Consumer sends its own name to the Producer as part of the negotiation
for Synchronous Replication.
A Producer needs to know the Consumer's name to identify the source of
received SeqnoAck messages.

Temporarily, with this patch the Consumer sends the connection name in
place of the Consumer's name. That is because the Consumer's name is
supposed to be given by ns_server at DCP_OPEN, but we still miss that
feature.
Note that the connection name contains the Consumer's name, so that will
be unique at Producer.

Change-Id: I92aefab2e2eafab3b36971ac9da438096ad89de3
Reviewed-on: http://review.couchbase.org/104133
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>