Refactor: Use a single packet validator
The packet validator consists of an array of 256 function pointers,
where each entry provides the validator for the opcode with the
value of the index into the array. This results in a memory
footprint of 2k. Given that the validator doesn't keep any
internal state there is no point of adding a 2k payload
to the bucket metadata (The pre-allocated bucket array is
set to 100 entries).

Change-Id: If89e4d0fdb0514f81e98b34ccbe6192e0a2a2be1
Reviewed-on: http://review.couchbase.org/104186
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>