MB-32107: Reduce false sharing of cache lines in DcpProducer
Resolve the performance regression caused by changing the size of
ConnHandler and subsequently the DcpProducer. The change to the size of
ConnHandler shuffled DcpProducer class members around which caused the 
unordered_map in StreamsMap to share a cache line with members in the
lock, and members in the lock of the StreamsMap to share a cache line
with following atomic variables in the DcpProducer. Resolve this issue
by cache line padding the mutex in the StreamsMap.

Perf testing was done against build 1553 and build 1558 of MadHatter.

Average results for runs:
1553 - 1973894 ops/s (8 runs)
1558 - 1923984 ops/s (8 runs)
Regressing Commit - 1938150 (5 runs)
        (http://review.couchbase.org/#/c/101191/ - based on 1553)
This Commit - 1972524 (9 runs)
        (based on 1553 + regressing commit)

Change-Id: I5d7b54e0a5f3348b7565a663993d56bc2165c300
Reviewed-on: http://review.couchbase.org/104193
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>