MB-32835: Fix stored value equivalence operator deletionSource handling
Fix an issue where (ordered) stored value's equivalence operator would
fail if the deletionSource did not match, but the value was not deleted.
In this case, the deletionSource should not be checked as it is updated
on a values deletion/expiration.
Also added a test for this issue, which requires the use of
a friend linked test fixture (StoredValueProtectedTest) to access the
protected deletionSource without deleting the stored value.

Change-Id: I2a4ae8fd9db1a416a88f45d41a4f11fbafb99cb3
Reviewed-on: http://review.couchbase.org/104199
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>