MB-32431: Remove costly checks from TaskQueue::_moveReadyTasks()
As part of the fixes for MB-32181, extra sanity checks were added to
TaskQueue::_moveReadyTasks() to check the futureQueue's
invariants. These were guarded under CB_DEVELOPMENT_ASSERTS so will be
disabled when we branch for release; however they are unfortunately
very costly - a 10% regresssion in replicateTo=1 throughput was
observed when this change was made.

As such, remove the call to futureQueue.assertInvariants - we haven't
actually seen it fire since was added ~2 months ago, so it's probalby
not worth the cost of the check anyway.

Change-Id: Ie2f931dd7f9fed094379415ce646c8986ef41f72
Reviewed-on: http://review.couchbase.org/104208
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>