MB-32107: Reduce cache line contention by moving DcpProducer acquisition
Currently the cache line that holds the control block of the DcpProducer
is one of the top three most contentious when running the daily
kv/kv_max_ops_writes.test. This is because in
ActiveStream::notifyStreamReady we lock the weak_ptr to the DcpProducer
before checking if we have any items to send. We don't need the
DcpProducer to check if we have any items to send, so swap the ordering
of these checks.

With this commit, there is no contention (hit modifieds) in the cahce
line containing the control block of the DcpProducer.

Perf testing of this commit was done against a build based on MadHatter
build 1553 + the performance regressing commit identified in MB-32107 +
the fix for the performance regressing commit as perf averages were
already known.

Average results for runs:
1553 - 1973894 ops/s (8 runs)
Build without this commit - 1972524 (9 runs)
        (MH - 1553 + regressing commit + fix for regressing commit)
Build with this commit - 1984044.625 (8 runs)
        (MH - 1553 + regressing commit + fix for regressing commit)

Change-Id: Ib0c65100f8de52d2e2a06b4666d8d14ddb7ec1bc
Reviewed-on: http://review.couchbase.org/104255
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>