Refactor: Modify connection limit handling
Modified the check in the packet validation phase to only
look at unathenticated connections. Whenever we fill up the
"user" available section of the connections we start refusing
clients to run other commands than the commands used in the
authentication phase.

As part of authentication we perform an extra check to see
the client authenticate as an internal user or not (and check
the appropriate limit).

Change-Id: Id90841392673b1f50b09bd17ee91cbeacd98216f
Reviewed-on: http://review.couchbase.org/104317
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>