MB-32807 [SR]: Enable Multiple-Replicas
With this patch we switch on Multiple-Replicas for Durability.
Two main changes here:

1) We remove any hard-coded node name in ReplicationChain. So now a
chain can be set only by ns_server through set-vbucket-state.

2) Given that ns_server doesn't pass the topology information yet, we
temporarily update the Replication Chain at master when the Producer
receives the 'consumer_name' via DcpControl.
Note that this is just a workaround, we'll remove it as soon as
ns_server provides the topology via set-vbucket-state.

Change-Id: I8413824adf62f5bcca5fca3f7bc91ea8875ab34a
Reviewed-on: http://review.couchbase.org/104101
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>