MB-32887: Make Eph Collections eraser delete deleted items
The collections eraser in ephemeral does not delete deleted items
when it iterates on the seqlist because fetchValidValue is called
with WantDeletedItems::No. This causes a heap use after free on
destruction of the hash table because we have removed this item from
the seqlist but it stil exists in the hash table.

Change the fetchValidValue call to fetch deleted items too, but only
erase values that are not system events. The system event for the end
of a collection will be released from the HTTombstoneCleaner and
marked as stale in the seqlist. The stale flag will then trigger the
removal from the seqlist by the StaleItemRemover.

Change-Id: I3b08cb22851f31c71f3c0774d7e430e21f04db3f
Reviewed-on: http://review.couchbase.org/104390
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>