MB-32902: Fix crash when calling FetchAllKeysTask
FetchAllKeys uses the new std::function version of AddResponseFn for
its callback. A copy of this callback is taken into the created
FetchAllKeysTask, which then runs on a background thread and from
there calls AddResponseFn.

Given the AddResponseFn callback is invoked on a background thead, is
runs _after_ the frontend thread call-stack has unwound - specifically
past the call to bucket_unknown_command() which is where the on-stack
AddResponseFn instance is created - specifically
process_bin_unknown_packet(). As a consequence the copied
FetchAllKeysTask::response member (which is actually an instance of
ExitGuard) contains a reference to a now-deleted AddResponseFn
(mcbp_response_handler). Calling this results in a crash (executing an
invalid std::function).

Fix this by creating a static instance of AddResponseFn and passing
that to ep-engine. This cannot go out of scope and hence solves the
problem.

stats_context.cc is vulnerable to the same issue (for append_stats),
so also apply the same fix pre-emptively there.

Change-Id: I6da95cfc454fb3fbc9d0df2488caf20c39804265
Reviewed-on: http://review.couchbase.org/104454
Reviewed-by: Daniel Owen <owend@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>