MB-31810: Track the persisted high seqno of system events
Currently, we do not track the high seqno of collection events
in the collections persisted high seqno. Track the high seqno of
these events by decoding the collection ID from the system event
key, and updating the appropriate collection's persisted high
seqno.

There exists a special case that must be dealt with. If an active
node has a collection tombstone, and is streaming to a replica that
has no knowledge of the collection, then it is possible to not find
the manifest entry corresponding to the collection (i.e. a fresh
replica). In this case, we should not throw an exception if we
cannot find the collection, we should just continue.

Change-Id: I6576aedda34e50739ea8b0f8bfcb8d42433cc983
Reviewed-on: http://review.couchbase.org/104466
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>