MB-32874: KVStore scan no longer needs CacheLookup help
The CacheLookup object held a reference to a manifest (which is
constructed from the persisted collection state) and then used that
manifest to test keys for being dropped.

As we move towards KVStore having more ownership of the persisted meta
there is no need for this complication, the implementation of KVStore
is changed so that it makes the decision about a key before invoking
the callback - thus CacheLookup is only called for 'live' keys.

A further fix is that system event keys are never checked for liveness
because we don't hide tombstones of collection/scope markers, they are
always passed up to the scan caller (i.e. DCP) for consideration within
their own use-case.

Change-Id: I985d9a1d12665ce448b6e22caa2ed2c4e75ec406
Reviewed-on: http://review.couchbase.org/104575
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>