MB-32937 [SR]: Allow undefined nodes in replication chains
A replication chain is permitted to have any of the replica nodes
marked as 'undefined' - the bucket has this many replicas, but nodes
have not yet been assigned to them (for example pending rebalance).

Update the handling of the topology passed via set_vbucket_state to
permit replaca nodes to be undefined - indicated by a Null type in the
JSON.

Note that the active node (the first element in each chain) is _not_
permitted to be undefined (Null).

Change-Id: Ic1faaf1faff25c128fb746a03e02969e6e4e45b7
Reviewed-on: http://review.couchbase.org/104626
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>