VBNotifyCtx: Remove Monotonic<> from bySeqno
While at the VBucket level bySeqno should never go backwards and hence
logically /is/ monotonic, in the typical use of VBNotifyCtx we would
default-construct an instance and then assign the result to it:

    VBNotifyCtx notifyCtx;
    std::tie(v, status, notifyCtx) =
        updateStoredValue(hbl, *v, itm, queueItmCtx)

In this scernario it's possible that the returned notifyCtx from
updateStoredValue is also default-constructed (e.g. if there was a
non-success status). As such it's not valid to require
VBNotifyCtx::bySeqno to be updated monotonically.

Change-Id: Ie1bb10d101f9d9bfc33366391060e73fb9a676b5
Reviewed-on: http://review.couchbase.org/104648
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>