MB-32591: Re-Introduce ChkPtQueueIterator
The initial patch had a performance issue due to performing a "deep"
compare of the CheckPointIterator container objects.

The effected code was:

auto operator==(CheckpointIterator ci) {
return (ci.iter == iter && ci.container.get() == container.get());
}

when should have been:

auto operator==(CheckpointIterator ci) {
return (ci.iter == iter && &(ci.container.get()) == &(container.get()));
}

This has been addressed and a corresponding performance benchmark has
been added.

Also in the original patch it did not convert the getLowSeqno and
getHighSeqno methods to using using the new iterator.  This has been
addressed in this patch.

-- Original commit message --

To help reduce memory usage we want to be able to remove items from the
checkpoint queue where all cursors are to the right (i.e. the items have
been processed) and hence don't need to be held in the queue.

The first step in this is to wrap the checkpoint queue iterator, (an
iterator on a stl::list) in a special iterator that automatically
skips passed elements that contain null pointers.

Change-Id: Ib68eedf3be064a5e6cc63cf7eaa2f8516c9de3dc
Reviewed-on: http://review.couchbase.org/104658
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>